Update Flex plugin version
