--- flex-plugin/public/appConfig.example.js
@@ -1,20 +1,0 @@
-// your account sid
-var accountSid = 'accountSid';
-
-// set to /plugins.json for local dev
-// set to /plugins.local.build.json for testing your build
-// set to "" for the default live plugin loader
-var pluginServiceUrl = '/plugins.json';
-
-var appConfig = {
-  pluginService: {
-    enabled: true,
-    url: pluginServiceUrl,
-  },
-  sso: {
-    accountSid: accountSid,
-  },
-  ytica: false,
-  logLevel: 'debug',
-  showSupervisorDesktopView: true,
-};
