--- flex-plugin/public/plugins.json
@@ -1,13 +1,0 @@
-[
-  {
-    "name": "plugin-sms-media",
-    "version": "0.0.0",
-    "class": "SmsMediaPlugin",
-    "requires": [
-      {
-        "@twilio/flex-ui": "^1"
-      }
-    ],
-    "src": "http://localhost:8080/plugin-sms-media.js"
-  }
-]
