--- flex-plugin/public/plugins.local.build.json
@@ -1,13 +1,0 @@
-[
-  {
-    "name": "SmsMediaPlugin",
-    "version": "0.0.0",
-    "class": "SmsMediaPlugin",
-    "requires": [
-      {
-        "@twilio/flex-ui": "^1"
-      }
-    ],
-    "src": "http://127.0.0.1:8085"
-  }
-]
