+++ plugin-message-media/jest.config.js
@@ -1,0 +1,8 @@
+module.exports = (config, { isProd, isDev, isTest }) => {
+  /**
+   * Customize the Jest by modifying the config object.
+   * Consult https://jestjs.io/docs/en/configuration for more information.
+   */
+
+  return config;
+}
