+++ plugin-message-media/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "plugin-message-media",
+  "version": "0.1.0",
+  "private": true,
+  "scripts": {
+    "postinstall": "flex-plugin pre-script-check"
+  },
+  "dependencies": {
+    "@material-ui/core": "^3.9.4",
+    "react": "16.5.2",
+    "react-dom": "16.5.2",
+    "flex-plugin-scripts": "4.3.19-beta.0"
+  },
+  "devDependencies": {
+    "@twilio/flex-ui": "^1"
+  }
+}
