+++ plugin-message-media/public/appConfig.example.js
@@ -1,0 +1,14 @@
+var accountSid = ''
+
+var appConfig = {
+  pluginService: {
+    enabled: true,
+    url: '/plugins',
+  },
+  sso: {
+    accountSid: accountSid
+  },
+  ytica: false,
+  logLevel: 'info',
+  showSupervisorDesktopView: true,
+};
