+++ plugin-message-media/src/SmsMediaPlugin.js
