+++ plugin-message-media/src/components/BubbleMessageWrapper/BubbleMessageWrapper.jsx
