+++ plugin-message-media/src/components/ImageModal/ImageModal.jsx
