+++ plugin-message-media/src/components/MediaMessage/MediaMessage.Styles.js
