+++ plugin-message-media/src/components/MediaMessage/MediaMessage.jsx
