+++ plugin-message-media/src/components/SendMediaComponent/SendMediaComponent.Styles.js
