+++ plugin-message-media/src/components/SendMediaComponent/SendMediaComponent.jsx
