+++ plugin-message-media/webpack.config.js
@@ -1,0 +1,8 @@
+module.exports = (config, { isProd, isDev, isTest }) => {
+  /**
+   * Customize the webpack by modifying the config object.
+   * Consult https://webpack.js.org/configuration for more information
+   */
+
+  return config;
+}
