+++ plugin-message-media/webpack.dev.js
@@ -1,0 +1,8 @@
+module.exports = (config, { isProd, isDev, isTest }) => {
+  /**
+   * Customize the webpack dev-server by modifying the config object.
+   * Consult https://webpack.js.org/configuration/dev-server for more information.
+   */
+
+  return config;
+}
